Add test, refactor properties
