+++ tests/test_persistence_experiment.py
@@ -1,5 +1,8 @@
 import os
 
+import joblib
+import pytest
+
 import pycaret.anomaly
 import pycaret.classification
 import pycaret.clustering
@@ -13,7 +16,17 @@ from pycaret.regression import RegressionExperiment
 from pycaret.time_series import TSForecastingExperiment
 
 
-def test_anomaly_persistence(tmpdir):
+def check_experiment_equality(exp, new_exp):
+    for key, value in exp.variables.items():
+        try:
+            assert value == new_exp.variables[key]
+        except Exception:
+            # For numpy arrays
+            assert joblib.hash(value) == joblib.hash(new_exp.variables[key])
+
+
+@pytest.mark.parametrize("preprocess_data", (True, False))
+def test_anomaly_persistence(tmpdir, preprocess_data):
     data = pycaret.datasets.get_data("anomaly")
     exp = AnomalyExperiment()
     exp.setup(
@@ -24,22 +37,19 @@ def test_anomaly_persistence(tmpdir):
         n_jobs=1,
     )
     exp_path = os.path.join(tmpdir, "exp.pkl")
-    exp.save_config(exp_path)
+    exp.save_experiment(exp_path)
 
-    new_exp = AnomalyExperiment()
-    new_exp.setup(
-        data,
-        html=False,
-        normalize=False,
-        session_id=123,
-        n_jobs=1,
+    new_exp = AnomalyExperiment.load_experiment(
+        exp_path,
+        data=data if preprocess_data else exp.data,
+        preprocess_data=preprocess_data,
     )
-    assert "normalize" not in new_exp.pipeline.named_steps
-    new_exp.load_config(exp_path)
     assert "normalize" in new_exp.pipeline.named_steps
+    check_experiment_equality(exp, new_exp)
 
 
-def test_clustering_persistence(tmpdir):
+@pytest.mark.parametrize("preprocess_data", (True, False))
+def test_clustering_persistence(tmpdir, preprocess_data):
     data = pycaret.datasets.get_data("jewellery")
     exp = ClusteringExperiment()
     exp.setup(
@@ -50,22 +60,19 @@ def test_clustering_persistence(tmpdir):
         n_jobs=1,
     )
     exp_path = os.path.join(tmpdir, "exp.pkl")
-    exp.save_config(exp_path)
+    exp.save_experiment(exp_path)
 
-    new_exp = ClusteringExperiment()
-    new_exp.setup(
-        data,
-        normalize=False,
-        html=False,
-        session_id=123,
-        n_jobs=1,
+    new_exp = ClusteringExperiment.load_experiment(
+        exp_path,
+        data=data if preprocess_data else exp.data,
+        preprocess_data=preprocess_data,
     )
-    assert "normalize" not in new_exp.pipeline.named_steps
-    new_exp.load_config(exp_path)
     assert "normalize" in new_exp.pipeline.named_steps
+    check_experiment_equality(exp, new_exp)
 
 
-def test_classification_persistence(tmpdir):
+@pytest.mark.parametrize("preprocess_data", (True, False))
+def test_classification_persistence(tmpdir, preprocess_data):
     data = pycaret.datasets.get_data("juice")
     exp = ClassificationExperiment()
     exp.setup(
@@ -77,23 +84,19 @@ def test_classification_persistence(tmpdir):
         n_jobs=1,
     )
     exp_path = os.path.join(tmpdir, "exp.pkl")
-    exp.save_config(exp_path)
+    exp.save_experiment(exp_path)
 
-    new_exp = ClassificationExperiment()
-    new_exp.setup(
-        data,
-        target="Purchase",
-        normalize=False,
-        html=False,
-        session_id=123,
-        n_jobs=1,
+    new_exp = ClassificationExperiment.load_experiment(
+        exp_path,
+        data=data if preprocess_data else exp.data,
+        preprocess_data=preprocess_data,
     )
-    assert "normalize" not in new_exp.pipeline.named_steps
-    new_exp.load_config(exp_path)
     assert "normalize" in new_exp.pipeline.named_steps
+    check_experiment_equality(exp, new_exp)
 
 
-def test_regression_persistence(tmpdir):
+@pytest.mark.parametrize("preprocess_data", (True, False))
+def test_regression_persistence(tmpdir, preprocess_data):
     data = pycaret.datasets.get_data("boston")
     exp = RegressionExperiment()
     exp.setup(
@@ -105,23 +108,19 @@ def test_regression_persistence(tmpdir):
         n_jobs=1,
     )
     exp_path = os.path.join(tmpdir, "exp.pkl")
-    exp.save_config(exp_path)
+    exp.save_experiment(exp_path)
 
-    new_exp = RegressionExperiment()
-    new_exp.setup(
-        data,
-        target="medv",
-        normalize=False,
-        html=False,
-        session_id=123,
-        n_jobs=1,
+    new_exp = RegressionExperiment.load_experiment(
+        exp_path,
+        data=data if preprocess_data else exp.data,
+        preprocess_data=preprocess_data,
     )
-    assert "normalize" not in new_exp.pipeline.named_steps
-    new_exp.load_config(exp_path)
     assert "normalize" in new_exp.pipeline.named_steps
+    check_experiment_equality(exp, new_exp)
 
 
-def test_time_series_persistence(tmpdir, load_pos_and_neg_data):
+@pytest.mark.parametrize("preprocess_data", (True, False))
+def test_time_series_persistence(tmpdir, load_pos_and_neg_data, preprocess_data):
     data = load_pos_and_neg_data
     exp = TSForecastingExperiment()
     exp.setup(
@@ -132,20 +131,22 @@ def test_time_series_persistence(tmpdir, load_pos_and_neg_data):
         n_jobs=1,
     )
     exp_path = os.path.join(tmpdir, "exp.pkl")
-    exp.save_config(exp_path)
+    exp.save_experiment(exp_path)
 
-    new_exp = TSForecastingExperiment()
-    new_exp.setup(
-        data,
-        html=False,
-        session_id=123,
-        n_jobs=1,
+    new_exp = TSForecastingExperiment.load_experiment(
+        exp_path,
+        data=data if preprocess_data else exp.data,
+        preprocess_data=preprocess_data,
     )
-    assert (
-        "transformer_target"
-        not in new_exp.pipeline.named_steps["forecaster"].named_steps
-    )
-    new_exp.load_config(exp_path)
-    assert (
-        "transformer_target" in new_exp.pipeline.named_steps["forecaster"].named_steps
-    )
+    # check experiment equality is not working for TS due to sktime,
+    # so we simply compare results
+    model = exp.create_model("ets")
+    results = exp.pull()
+    preds = exp.predict_model(model)
+
+    new_model = new_exp.create_model("ets")
+    new_results = new_exp.pull()
+    new_preds = new_exp.predict_model(new_model)
+
+    assert preds.equals(new_preds)
+    assert results.equals(new_results)
